Bug fix: access error message due to manage_client_links permission; added main lead office list table

+++ app/views/units/show.blade.php
@@ -55,7 +55,7 @@
 	<h3>Active clients linked to this unit</h3>
 </div>
 
-@if($user->hasRole('Administrator') && count($clients) > 0)
+ @if(count($clients) > 0)
 	<section class="index-table-container">
 		<div class="row">
 			<div class="col-12">
@@ -66,7 +66,9 @@
 							<td width="15%" class="content-center hide-s">Sector</td>
 							<td width="15%" class="content-center hide-m">Type</td>
 							<td width="10%" class="content-center hide-m">Service</td>
-							<td colspan="3">Actions</td>
+							@if($user->hasRole('Administrator'))
+								<td colspan="3">Actions</td>
+							@endif
 						</tr>
 					</thead>
 					<tbody>
@@ -76,21 +78,23 @@
 								<td class="content-center hide-s">{{ $client->sector()->pluck('name') }}</td>
 								<td class="content-center hide-m">{{ $client->type()->pluck('name') }}</td>
 								<td class="content-center hide-m">{{ $client->service()->pluck('name') }}</td>
-								<td class="actions content-right">
-									{{ Form::open(['url' => 'clients/' . $client->id . '/archive', 'method' => 'get']) }}
-										<button type="submit" class="primary" title="Add an archive entry for this client"><i class="fa fa-folder"></i></button>
-									{{ Form::close() }}
-								</td>
-								<td class="actions content-right">
-									{{ Form::open(['route' => array('clients.edit', $client->id), 'method' => 'get']) }}
-										<button type="submit" class="primary" title="Edit this client"><i class="fa fa-pencil"></i></button>
-									{{ Form::close() }}
-								</td>
-								<td class="actions content-right">
-									{{ Form::open(['route' => array('clients.destroy', $client->id), 'method' => 'delete']) }}
-										<button type="submit" class="red-but delete-row" data-resource-type="sector" title="Delete this client"><i class="fa fa-times"></i></button>
-									{{ Form::close() }}
-								</td>
+								@if($user->hasRole('Administrator'))
+									<td class="actions content-right">
+										{{ Form::open(['url' => 'clients/' . $client->id . '/archive', 'method' => 'get']) }}
+											<button type="submit" class="primary" title="Add an archive entry for this client"><i class="fa fa-folder"></i></button>
+										{{ Form::close() }}
+									</td>
+									<td class="actions content-right">
+										{{ Form::open(['route' => array('clients.edit', $client->id), 'method' => 'get']) }}
+											<button type="submit" class="primary" title="Edit this client"><i class="fa fa-pencil"></i></button>
+										{{ Form::close() }}
+									</td>
+									<td class="actions content-right">
+										{{ Form::open(['route' => array('clients.destroy', $client->id), 'method' => 'delete']) }}
+											<button type="submit" class="red-but delete-row" data-resource-type="sector" title="Delete this client"><i class="fa fa-times"></i></button>
+										{{ Form::close() }}
+									</td>
+								@endif
 							</tr>
 						@endforeach
 					</tbody>
